Readme file / gitignore file added (start project)
Added a .gitignore locally with file explorer with content that might be added automatically (for example. __pycache__/), content that might be user/machine specific (for example. .env) or sensitive data (for example. settings.py). And a README.md file with a layout that will be configured at a later stage.